Temporal Naval Fix
Starkku at the same time of this implementation published a PR for solving this bug "in the right way" but meanwhile I'll use this "dirty" code that does almost the same.

Also I fixed all the bugs related to the Teams categorization, now all the AI teams go to the right caegory depending the members of the Taskforce.

Added 2 new boolean tags for overwriting the categorization of some units:
- ConsideredNaval
- ConsideredVehicle