First commit
Currently is a workaround that needs the CnCNet 5 library to work and allows the use of campaign missions outside the mix files using SkipMapSelect tag and NextScenario & AltNextScenario tags for pointing to the next mission.
MISSIONMD.ini is used for the loading screens & other small details, being the relative path of the map the section name. For example [Maps/Missions/RA2/alld99.map]